feat: Token-based authentication & user administration

+++ part4/controllers/blogs.js
@@ -1,17 +1,31 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 // Create a blog
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    url: body.url,
+    title: body.title,
+    author: body.author,
+    user: user._id
+  })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
 // Read all blogs
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -19,6 +33,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.put('/:id', async (request, response) => {
   const { likes } = request.body
   const blog = await Blog.findById(request.params.id)
+
   if (!blog) {
     return response.status(404).end()
   }
@@ -28,9 +43,19 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 // Delete a blog
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
+
 
 module.exports = blogsRouter
